Clean up some unused functions
