+++ ADLER/ADLERGUI.py
@@ -57,7 +57,7 @@ from ExtendGUI import LogBox
 from FileFinder import PeaxisDataModel
 from ExperimentTree import TableView
 
-ADLER_VERSION_STRING = "4.1 from 27.10.2022"
+ADLER_VERSION_STRING = "4.2 from 09.04.2023"
 
 mpl_scale = 1.0
 mpl_figure_scale = 1.0
@@ -191,27 +191,6 @@ from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as Navigatio
 GlobFont = QFont('Sans Serif', int(12*font_scale))
 
 oldval = 0.0
-
-#### filesystem monitoring part
-def FindUnprocessedFiles(fpath):
-    infiles, outfiles = [], []
-    # with os.scandir(fpath) as it:
-    for entry in os.scandir(fpath):
-        if entry.is_file():
-            tokens = entry.name.split('.')
-            name, extension = '.'.join(tokens[:-1]), tokens[-1]
-            if extension == 'sif':
-                infiles.append(name)
-            elif extension == 'asc':
-                if name[-3:] == '_1D':
-                    outfiles.append(name[:-3])
-                else:
-                    outfiles.append(name[:-3])
-    unp_files = []
-    for fnam in infiles:
-        if not fnam in outfiles:
-            unp_files.append(fnam)
-    return unp_files
 
 #### GUI part
 
@@ -361,12 +340,6 @@ class BetterDict():
         for kk in input_keys:
             self.dict[kk[0]][kk[1]] = copy.deepcopy(newvalues[kk[0]][kk[1]])
 
-class QHLine(QFrame):
-    def __init__(self):
-        super().__init__()
-        self.setFrameShape(QFrame.HLine)
-        self.setFrameShadow(QFrame.Sunken)
-
 class InterfaceTabbed:
     def __init__(self, master, size, param,  appinstance = None, pathdict = {}):
         self.master = master
@@ -430,7 +403,6 @@ class InterfaceTabbed:
         # self.master.resize(size[0], size[1])
         self.master.setCentralWidget(self.tabbar)
         self.master.setFont(GlobFont)
-        # self.master.resized.connect(self.on_resize)
         self.load_last_params()
         # self.save_last_params()
         self.logger("MAX_THREADS set to " + str(MAX_THREADS))
@@ -442,8 +414,6 @@ class InterfaceTabbed:
         self.master.destroyed.connect(self.confthread.quit)
         self.conf_writer.moveToThread(self.confthread)
         self.confthread.start()
-    def on_resize(self):
-        self.master.resize(self.master.sizeHint())
     def MakeFileTab(self):
         tab1 = QWidget(self.tabbar)
         # here comes the new layout
